+++ client/app/form-builder/form-builder.controller.js
@@ -105,7 +105,7 @@ angular.module('formBuilder')
             return;
           }
 
-          var version = LForms.Util.detectFHIRVersion(importedData);
+          var version = LForms.Util.detectFHIRVersion(importedData) || LForms.Util.guessFHIRVersion(importedData);
           version = !version ? 'lforms' : version;
           switch(version) {
             case 'STU3':
@@ -169,13 +169,15 @@ angular.module('formBuilder')
           $scope.termsOfUseAccepted = 'no';
         });
       };
+
+
       /**
        * Setter - Safe way to assign from child scopes.
        *
        * @param node
        */
       $scope.setSelectedNode = function(node) {
-        if($scope.selectedNode !== node && node) {
+        if($scope.selectedNode !== node && node && node.lfData && node.lfData.basic.code === 'basicItemLevelFields') {
           Def.Autocompleter.screenReaderLog('A new item '+
             $scope.getItem(node.lfData, 'question').value +
             ' is loaded into item builder panel');
@@ -254,7 +256,7 @@ angular.module('formBuilder')
             var content = $scope.previewSource[answer.format];
             var blob = new Blob([content], {type: 'application/json;charset=utf-8'});
             var formName = $scope.formBuilderData.treeData[0].lfData.basic.itemHash['/name/1'].value;
-            var exportFileName = formName ?  formName.replace(/\s/g, '-') : 'new-form';
+            var exportFileName = formName ?  formName.replace(/\s/g, '-') : 'form';
 
             // Use hidden anchor to do file download.
             var downloadLink = document.getElementById('exportAnchor');
@@ -551,7 +553,7 @@ angular.module('formBuilder')
         var indexInfo = dataConstants.INITIAL_FIELD_INDICES[fieldName];
 
         if(indexInfo) {
-          ret = lfData[indexInfo.category].items[indexInfo.index];
+          ret = formBuilderService.getFormBuilderField(lfData[indexInfo.category].items, fieldName);
         }
         else {
           var key = '/'+fieldName+'/1';
@@ -593,9 +595,8 @@ angular.module('formBuilder')
       $scope.isHeaderItem = function (lfData) {
         var ret = false;
         if(formBuilderService.isNodeFbLfItem(lfData)) {
-          ret = lfData[dataConstants.INITIAL_FIELD_INDICES['header'].category]
-            .items[dataConstants.INITIAL_FIELD_INDICES['header'].index]
-            .value.code;
+          var indexInfo = dataConstants.INITIAL_FIELD_INDICES['header'];
+          ret = formBuilderService.getFormBuilderField(lfData[indexInfo.category].items, 'header').value.code;
         }
         return ret;
       };